feat: Update package structure
